Revert change from % to ^
I think my change of the spec from % to ^ was premature,
partially because I now mildly prefer % again,
but mostly because it’s what we presented to plenary
on August and because it’s what Babel currently supports.

This is a tentative change and has no bearing on what the actual final topic token will be.

As per https://github.com/tc39/proposal-pipeline-operator/issues/91#issuecomment-926357211.

See also https://github.com/tc39/proposal-pipeline-operator/issues/91#issuecomment-911851451,
which I think was now premature.