Revert change from % to ^ (#250)
As per https://github.com/tc39/proposal-pipeline-operator/issues/91#issuecomment-926357211, I think my [change of the spec from `%` to `^`](https://github.com/tc39/proposal-pipeline-operator/issues/91#issuecomment-911851451) was premature, partially because I now mildly prefer `%` again, but mostly because it’s what we presented to plenary on August and because it’s what Babel currently supports (see babel/babel#13749).

This is a tentative change and has no bearing on what the actual final topic token will be. It’s just a swap to the explainer and spec back to what we presented to the Committee plenary on August.

This will be almost certainly be the last time I switch the topic token until we hold our TC39 incubator sessions for bikeshedding the token in early November.